final refactor on db mock mem object
