+++ cushion/persist/base.py
@@ -1,0 +1,5 @@
+
+
+class BaseConnection(object):
+    """ the base connection type.. python needs interfaces """
+    pass
